+++ crates/model/src/python/orders/market_if_touched.rs
@@ -125,8 +125,8 @@ impl MarketIfTouchedOrder {
 
     #[staticmethod]
     #[pyo3(name = "create")]
-    fn py_create(init: OrderInitialized) -> Self {
-        Self::from(init)
+    fn py_create(init: OrderInitialized) -> PyResult<Self> {
+        Self::try_from(init).map_err(to_pyvalue_err)
     }
 
     #[staticmethod]